requirements/production.txt
# Production Only
gunicorn

# Production and Development
Django
psycopg2
django-mptt
-e git+https://github.com/zyga/django-pagination.git@master#egg=linaro-django-pagination
django-filter
webstack-django-sorting
djangorestframework
Pillow
django-haystack
pyelasticsearch
django-froala-editor
# -e git+https://github.com/helber/django-dbsettings.git@master#egg=django-dbsettings
django-modeltranslation
-e git+https://github.com/awecode/njango.git@master#egg=njango
sorl-thumbnail==11.12.1b
django-redis
jsonfield
openpyxl
WeasyPrint
requests
django-admin-sortable2
django-solo

django-import-export==0.5.1